third_party/WebKit/LayoutTests/web-animations-api/player-unresolved-times.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
function assert_unresolved(value) {
  assert_equals(value, null);
}

test(function() {
  var player = document.documentElement.animate([], 100000);
  player.cancel();
  assert_unresolved(player.startTime);
  assert_unresolved(player.currentTime);
}, "unresolved startTime and currentTime");

</script>
